# feat: Add get_recent_changes tool with major performance optimizations