--- a/.claude/settings.local.json
+++ b/.claude/settings.local.json
@@ -1,9 +1,0 @@
-{
-  "permissions": {
-    "allow": [
-      "WebSearch"
-    ],
-    "deny": [],
-    "ask": []
-  }
-}