## Summary
- Add new `get_recent_changes` MCP tool for viewing recent commits with diffs
- Fix major performance issue in `read_blog_post` (99% improvement: 74s → 0.4s)
- Optimize test suite runtime by 98% (100s → 1.66s)

## Changes

### New Feature: get_recent_changes Tool
- View recent commits from the GitHub repository
- Support for time-based filtering (days parameter)
- Support for count-based filtering (commits parameter)
- Path filtering for specific directories
- Optional diff inclusion
- Parallel API requests with semaphore limiting (max 15 concurrent)

### Performance Optimizations
- **Fixed redirect checking in read_blog_post**: Now uses cached back-links.json instead of downloading all 200+ files
- **Hybrid testing strategy**: One real API test per tool as sanity check, mocked tests for edge cases
- **DRY improvements**: Added helper functions to eliminate duplicate code patterns

### Testing
- Comprehensive unit tests with both real and mocked GitHub API calls
- Full E2E test coverage for all parameter combinations
- Tests for error handling and edge cases

## Test Plan
- [x] Unit tests pass
- [x] E2E tests pass
- [x] Performance benchmarks show 98% improvement
- [x] All existing functionality preserved
- [x] Backward compatibility maintained

🤖 Generated with [Claude Code](https://claude.ai/code)